+++ internal/types/balance.go
@@ -5,7 +5,7 @@ import "time"
 type Balance struct {
 	Id        int       `json:"-"`
 	UserId    int       `json:"-"`
-	Sum       int       `json:"current"`
+	Sum       float64   `json:"current"`
 	CreatedAt time.Time `json:"-"`
 	UpdatedAt time.Time `json:"-"`
 }
